Options are saved and retrieved.

+++ core/src/com/mygdx/game/MyGdxGame.java
@@ -18,9 +18,9 @@ public class MyGdxGame extends Game {
 	public static final String ASSET_PATH = "asset" + File.separator;
 	
 	//used in the OptionsMenu
-	public static int DIFFICULTY;
+	public static float DIFFICULTY;
 	public static float RESOLUTION;
-	public static int VOLUME;
+	public static float VOLUME;
 	public static boolean LIGHTING;
 	public static boolean ANIMATION;
